@@ -14,7 +14,7 @@
             "php ./vendor/phpunit/phpunit/phpunit"
         ],
         "documentation": [
-            "php ./vendor/phpdocumentor/phpdocumentor/bin/phpdoc --title \"pickles2/px2agent v0.0.4 API Document\" -d \"./php/\" -t \"./docs/\""
+            "php ./vendor/phpdocumentor/phpdocumentor/bin/phpdoc --title \"pickles2/px2agent v0.1.0 API Document\" -d \"./php/\" -t \"./docs/\""
         ]
     } ,
     "minimum-stability": "dev",
@@ -22,15 +22,15 @@
     "repositories": [
     ],
     "require": {
-        "php": ">=5.4.0",
+        "php": ">=7.3.0",
         "ext-json": "*",
         "ext-mbstring": "*",
         "tomk79/filesystem": "~1.0",
         "tomk79/request": "~1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.1",
-        "phpdocumentor/phpdocumentor": "^2.9.0",
+        "phpunit/phpunit": "~9.5",
+        "phpdocumentor/phpdocumentor": "^3.0.0",
         "pickles2/px-fw-2.x": "~2.0"
     },
     "autoload": {
